+++ alacritty/src/config/terminal.rs
@@ -4,12 +4,14 @@ use toml::Value;
 use alacritty_config_derive::{ConfigDeserialize, SerdeReplace};
 use alacritty_terminal::term::Osc52;
 
-use crate::config::ui_config::StringVisitor;
+use crate::config::ui_config::{Program, StringVisitor};
 
-#[derive(ConfigDeserialize, Default, Copy, Clone, Debug, PartialEq)]
+#[derive(ConfigDeserialize, Default, Clone, Debug, PartialEq)]
 pub struct Terminal {
     /// OSC52 support mode.
     pub osc52: SerdeOsc52,
+    /// Path to a shell program to run on startup.
+    pub shell: Option<Program>,
 }
 
 #[derive(SerdeReplace, Default, Copy, Clone, Debug, PartialEq)]